+++ CHANGELOG.md
@@ -6,6 +6,12 @@ See DataLab [roadmap page](https://datalab-platform.com/en/contributing/roadmap.
 
 💥 New features and enhancements:
 
+* New image processing features:
+  * Added new "Gaussian frequency filter" feature:
+    * This feature allows to filter an image in the frequency domain using a Gaussian filter
+    * The feature is implemented in the `sigima.proc.image.freq_fft` function (requires Sigima v0.3)
+    * This closes [Issue #205](https://github.com/DataLab-Platform/DataLab/issues/205) - Add Gaussian frequency filter for images
+
 * Signal cursors:
   * Fixed [Issue #231](https://github.com/DataLab-Platform/DataLab/issues/231) - Add horizontal cursors to the Signal View (requires PlotPy v2.8)
   * Fixed [Issue #232](https://github.com/DataLab-Platform/DataLab/issues/232) - Add Δx measurement on curve stats tool (vertical cursors)